Add transformation abstraction
