+++ fol-ir/src/main/java/com/foxowlet/fol/ast/traverse/Transformation.java
@@ -1,0 +1,7 @@
+package com.foxowlet.fol.ast.traverse;
+
+import com.foxowlet.fol.ast.Node;
+
+public interface Transformation {
+    Node apply(Node node);
+}
